+++ include/Engine/Macros/Utils.hpp
@@ -19,14 +19,28 @@
 #define ENGINE_FUNC_SIG __PRETTY_FUNCTION__
 #endif
 
-constexpr inline bool is_ident_char(const char c) {
-  return (c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z') || (c >= '0' && c <= '9') || c == '_' || c == ':';
-}
-
-constexpr inline std::string_view className(std::string_view prettyFunction) {
+constexpr std::string_view className(const std::string_view prettyFunction, const bool fullyQualified) {
   const size_t colons = prettyFunction.rfind("::");
   if (colons == std::string_view::npos)
     return "::";
+
+  if (!fullyQualified) {
+    const size_t end = colons;
+    size_t start = end;
+    int angle_depth = 0;
+
+    while (start > 0) {
+      --start;
+      if (const char c = prettyFunction[start]; c == '>')
+        angle_depth++;
+      else if (c == '<')
+        angle_depth--;
+      else if (angle_depth == 0 && (c == ':' || c == ' '))
+        break;
+    }
+    ++start;
+    return prettyFunction.substr(start, end - start);
+  }
 
   const size_t end = colons;
   size_t start = end;
@@ -38,13 +52,15 @@ constexpr inline std::string_view className(std::string_view prettyFunction) {
       angle_depth++;
     else if (c == '<')
       angle_depth--;
-    else if (angle_depth == 0 && !is_ident_char(c))
+    else if (angle_depth == 0 &&
+             !((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z') || (c >= '0' && c <= '9') || c == '_' || c == ':'))
       break;
   }
   ++start;
   return prettyFunction.substr(start, end - start);
 }
 
-#define ENGINE_CLASS_NAME className(ENGINE_FUNC_SIG)
+#define ENGINE_CLASS_NAME className(ENGINE_FUNC_SIG, false)
+#define ENGINE_CLASS_NAME_FULLY_QUALIFIED className(ENGINE_FUNC_SIG, true)
 
 #endif //UTILS_H